Add CombatDisconnect
and rename AutoDisconnect
